+++ packages/alea-frontend/pages/api/nap/create-answer.ts
@@ -10,11 +10,14 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
   if (!checkIfPostOrSetError(req, res)) return;
   const userId = await getUserIdOrSetError(req, res);
   if (!userId) return;
-  const { answer, questionId, questionTitle, subProblemId } = req.body as CreateAnswerRequest;
+  const { questionId, questionTitle, subProblemId, courseId } = req.body as CreateAnswerRequest;
+  let { courseInstance, answer } = req.body as CreateAnswerRequest;
   if (!answer || !questionId || !questionTitle) res.status(422).end();
+  if (!courseInstance) courseInstance = 'current';
+  answer = answer.trim();
   const result = await executeAndEndSet500OnError(
-    `INSERT INTO Answer (questionId, userId, answer,questionTitle,subProblemId) VALUES (?,?,?,?,?)`,
-    [questionId, userId, answer, questionTitle, subProblemId],
+    `INSERT INTO Answer (questionId, userId, answer,questionTitle,subProblemId,courseId,courseInstance) VALUES (?,?,?,?,?,?)`,
+    [questionId, userId, answer, questionTitle, subProblemId, courseId,courseInstance],
     res
   );
   res.status(201).send({ id: result.insertId });
